feat: migrate language convention detection to LLM-based approach
- Replace programmatic pattern matching with LLM-based detection using Claude Sonnet 4.5
- Implement two-step reasoning approach:
  - Step 1: LLM gathers evidence (spelling patterns, word choices, grammar)
  - Step 2: LLM evaluates evidence and provides conclusion with confidence/consistency scores
- Change text truncation from 2000 characters to 2000 words for more consistent analysis
- Add evidence validation to handle empty arrays before Step 2
- Move extended UK English test (~2000 words) to integration test infrastructure
- Maintain backward compatibility with existing output schema
